+++ Editor/CreateClass.cs
@@ -6,16 +6,7 @@ namespace Foundation.Editor
 {
 	public class CreateClass
 	{
-		[MenuItem("Foundation/Patterns/Create Facade")]
-		private static void CreateFacade()
-		{
-			string path = "Assets/Facade.cs";
-			string body = "namespace Foundation.Patterns\n{\n	static class Facade\n	{\n	}\n}";
-
-			CreateFile(path, body);
-		}
-
-		private static void CreateFile(string path, string body)
+		public static void CreateFile(string path, string body)
 		{
 			Debug.Log("Creating Classfile: " + path);
 			if (File.Exists(path) == false)
@@ -23,6 +14,7 @@ namespace Foundation.Editor
 				using StreamWriter outfile = new StreamWriter(path);
 				outfile.Write(body);
 			}
+
 			AssetDatabase.Refresh();
 		}
 	}